+++ lib/screens/cart/components/order_review.dart
@@ -1,4 +1,5 @@
 import 'package:aishop/screens/cart/checkout_page.dart';
+import 'package:aishop/screens/cart/components/edit_cart.dart';
 import 'package:aishop/styles/theme.dart';
 import 'package:aishop/utils/cart.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
@@ -11,6 +12,7 @@ int count = 0;
 
 void updateCartTotal() async {
   double total = 0;
+  int qty = 0;
   await FirebaseFirestore.instance
       .collection('Users')
       .doc(FirebaseAuth.instance.currentUser!.uid)
@@ -18,8 +20,7 @@ void updateCartTotal() async {
       .get()
       .then((value) {
     value.docs.forEach((element) {
-      total +=
-          double.parse(element.data()['price']) * element.data()['quantity'];
+      total += double.parse(element.data()['total']);
     });
   });
   g = total;
@@ -51,7 +52,7 @@ class _OrderReviewState extends State<OrderReview> {
           primary: false,
           leading: Container(),
           title: Text(
-            'Order Review',
+            'Shopping Cart',
             style: TextStyle(
               fontSize: 25,
               color: lightblack,
@@ -80,8 +81,11 @@ class _OrderReviewState extends State<OrderReview> {
                     prodquantity: snapshot.data!.docs[index].get('quantity'),
                     proddescription:
                         snapshot.data!.docs[index].get('description'),
-                    prodprice: snapshot.data!.docs[index].get('price'),
+                    prodprice:
+                        snapshot.data!.docs[index].get('price').toString(),
                     prodindex: index,
+                    prodtotal:
+                        snapshot.data!.docs[index].get('total').toString(),
                   );
                 },
                 itemCount: snapshot.data!.docs.length,
@@ -96,258 +100,16 @@ class _OrderReviewState extends State<OrderReview> {
               Expanded(
                   child: ListTile(
                 title: new Text(
-                  "ORDER SUBTOTAL                      R $g\n                             ",
+                  " Cart Summary ",
                   style: TextStyle(color: white),
                 ),
                 subtitle: new Text(
-                  " \n TOTAL                                R $g",
+                  " \n TOTAL :                       R $g",
                   style: TextStyle(color: white, fontSize: 19.0),
                 ),
               ))
             ],
           ),
         ));
-  }
-}
-
-class SingleCartProduct extends StatefulWidget {
-  final prodname;
-  final prodpicture;
-  final prodprice;
-  final prodquantity;
-  final proddescription;
-  final cartid;
-  final prodindex;
-  final stockamt;
-
-  SingleCartProduct(
-      {this.prodname,
-      this.prodpicture,
-      this.prodprice,
-      this.prodquantity,
-      this.proddescription,
-      this.prodindex,
-      this.stockamt,
-      this.cartid});
-
-  @override
-  _SingleCartProductState createState() => _SingleCartProductState();
-}
-
-class _SingleCartProductState extends State<SingleCartProduct> {
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      color: Color(0xB7242424),
-      child: ListTile(
-        leading: new Image.network(
-          widget.prodpicture,
-          width: 70.0,
-          height: 70.0,
-        ),
-        title: new Text(widget.prodname, style: TextStyle(color: white)),
-        subtitle: new Column(
-          children: <Widget>[
-            new Row(
-              children: <Widget>[
-                Expanded(
-                  child: new Text(
-                    widget.proddescription,
-                    style: TextStyle(color: lightgrey),
-                    maxLines: 1,
-                  ),
-                )
-              ],
-            ),
-
-            // new Container(
-            //   alignment: Alignment.bottomLeft,
-            //   child: new Text(prodquantity,
-            //       style: TextStyle(color: Color(0xFFFDD835))),
-            // ),
-            SizedBox(
-              height: 10,
-            ),
-
-            new Row(
-              children: [
-                GestureDetector(
-                  child: Container(
-                    width: 20,
-                    height: 20,
-                    decoration: BoxDecoration(
-                        shape: BoxShape.circle,
-                        border: Border.all(color: Colors.yellow)),
-                    child: Center(
-                        child: Icon(
-                      Icons.remove,
-                      size: 10,
-                      color: Colors.yellowAccent,
-                    )),
-                  ),
-                  onTap: () async {
-                    if (widget.prodquantity > 1) {
-                      await FirebaseFirestore.instance
-                          .collection('Users')
-                          .doc(FirebaseAuth.instance.currentUser!.uid)
-                          .collection("Cart")
-                          .where('name', isEqualTo: widget.prodname)
-                          .get()
-                          .then((value) => value.docs.forEach((element) => {
-                                element.reference.update(
-                                    {"quantity": FieldValue.increment(-1)})
-                              }));
-
-                      setState(() {
-                        g -= double.parse(widget.prodprice);
-                        updateCartTotal();
-                        Navigator.pushReplacement(
-                            context,
-                            new MaterialPageRoute(
-                                builder: (context) => CheckOutPage()));
-                      });
-                    } else if (widget.prodquantity == 1) {
-                      Cart.removeFromCart(
-                          widget.cartid,
-                          widget.prodpicture,
-                          widget.proddescription,
-                          widget.prodname,
-                          widget.prodprice,
-                          widget.prodquantity,
-                          widget.stockamt);
-                      setState(() {
-                        g -= double.parse(widget.prodprice);
-                        updateCartTotal();
-                        Navigator.pushReplacement(
-                            context,
-                            new MaterialPageRoute(
-                                builder: (context) => CheckOutPage()));
-                      });
-                    }
-                  },
-                ),
-                SizedBox(
-                  height: 5,
-                  width: 10,
-                ),
-                Text(
-                  "${widget.prodquantity}",
-                  style: TextStyle(fontSize: 10, color: Colors.yellowAccent),
-                ),
-                SizedBox(
-                  height: 5,
-                  width: 10,
-                ),
-                GestureDetector(
-                  child: Container(
-                    width: 20,
-                    height: 20,
-                    decoration: BoxDecoration(
-                        shape: BoxShape.circle,
-                        border: Border.all(color: Colors.yellow)),
-                    child: Center(
-                        child: Icon(
-                      Icons.add,
-                      size: 10,
-                      color: Colors.yellowAccent,
-                    )),
-                  ),
-                  onTap: () async {
-                    await FirebaseFirestore.instance
-                        .collection('Users')
-                        .doc(FirebaseAuth.instance.currentUser!.uid)
-                        .collection("Cart")
-                        .where('name', isEqualTo: widget.prodname)
-                        .get()
-                        .then((value) => value.docs.forEach((element) => {
-                              element.reference
-                                  .update({"quantity": FieldValue.increment(1)})
-                            }));
-
-                    setState(() {
-                      g += double.parse(widget.prodprice);
-                      updateCartTotal();
-                      Navigator.pushReplacement(
-                          context,
-                          new MaterialPageRoute(
-                              builder: (context) => CheckOutPage()));
-                    });
-                  },
-                ),
-              ],
-            ),
-
-            SizedBox(
-              height: 5,
-            ),
-            new Container(
-              alignment: Alignment.bottomRight,
-              child: new Text(
-                  "R" +
-                      (double.parse(widget.prodprice) * widget.prodquantity)
-                          .toString(),
-                  style: TextStyle(color: Color(0xFFFDD835))),
-            ),
-
-            SizedBox(
-              height: 10,
-            ),
-
-            new Row(
-              children: <Widget>[
-                Expanded(
-                    child: Padding(
-                  padding: const EdgeInsets.fromLTRB(0.0, 0.0, 0.0, 0.0),
-                  child: Row(
-                    children: <Widget>[
-                      SizedBox(
-                        height: 0,
-                        width: 0,
-                      ),
-                      ElevatedButton.icon(
-                        onPressed: () {
-                          Cart.removeFromCart(
-                              widget.cartid,
-                              widget.prodpicture,
-                              widget.proddescription,
-                              widget.prodname,
-                              widget.prodprice,
-                              widget.prodquantity,
-                              widget.stockamt);
-                          setState(() {
-                            g -= double.parse(widget.prodprice) *
-                                widget.prodquantity;
-                            Navigator.pushReplacement(
-                                context,
-                                new MaterialPageRoute(
-                                    builder: (context) => CheckOutPage()));
-                            updateCartTotal();
-                          });
-                        },
-                        style: ElevatedButton.styleFrom(
-                          primary: Colors.grey,
-                        ),
-                        icon: Icon(
-                          Icons.delete_outline_rounded,
-                          size: 15,
-                          color: Colors.black,
-                        ),
-                        label: Text(
-                          "Remove",
-                          style: TextStyle(
-                              fontSize: 10.0,
-                              fontWeight: FontWeight.bold,
-                              color: Colors.yellowAccent),
-                        ),
-                      ),
-                    ],
-                  ),
-                ))
-              ],
-            ),
-          ],
-        ),
-      ),
-    );
   }
 }
